fix(diagnostics): keep backlog policy internal
