+++ src/logging/diagnostic-runtime.ts
@@ -3,6 +3,7 @@ import {
   areDiagnosticsEnabledForProcess,
   emitInternalDiagnosticEvent as emitDiagnosticEvent,
 } from "../infra/diagnostic-events.js";
+import { getDiagnosticSessionState, type SessionRef } from "./diagnostic-session-state.js";
 import { createSubsystemLogger } from "./subsystem.js";
 
 // Shared diagnostic logger and queue-activity event helpers.
@@ -25,6 +26,45 @@ export function getLastDiagnosticActivityAt(): number {
 /** Clears diagnostic activity state for tests. */
 export function resetDiagnosticActivityForTest(): void {
   lastActivityAt = 0;
+}
+
+type DiagnosticMessageQueueParams = SessionRef & {
+  channel?: string;
+  source: string;
+};
+
+/** Records queue activity while letting internal run owners distinguish steering from backlog. */
+export function logMessageQueuedWithBacklogPolicy(
+  params: DiagnosticMessageQueueParams,
+  countsTowardBacklog: boolean,
+): void {
+  if (!areDiagnosticsEnabledForProcess()) {
+    return;
+  }
+  const state = getDiagnosticSessionState(params);
+  if (countsTowardBacklog) {
+    state.queueDepth += 1;
+  }
+  state.lastActivity = Date.now();
+  state.generation = (state.generation ?? 0) + 1;
+  state.lastStuckWarnAgeMs = undefined;
+  state.lastLongRunningWarnAgeMs = undefined;
+  if (diag.isEnabled("debug")) {
+    diag.debug(
+      `message queued: sessionId=${state.sessionId ?? "unknown"} sessionKey=${
+        state.sessionKey ?? "unknown"
+      } source=${params.source} queueDepth=${state.queueDepth} sessionState=${state.state}`,
+    );
+  }
+  emitDiagnosticEvent({
+    type: "message.queued",
+    sessionId: state.sessionId,
+    sessionKey: state.sessionKey,
+    channel: params.channel,
+    source: params.source,
+    queueDepth: state.queueDepth,
+  });
+  markDiagnosticActivity();
 }
 
 /** Logs and emits a diagnostic event when work enters a serialized lane. */
